Factor accel calculations into calculation classes
