+++ grapher/grapher.csproj
@@ -47,6 +47,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Models\Calculations\AccelCalculator.cs" />
+    <Compile Include="Models\Calculations\AccelData.cs" />
     <Compile Include="Models\Charts\AccelCharts.cs" />
     <Compile Include="Models\AccelGUI.cs" />
     <Compile Include="Models\Options\AccelOptions.cs" />
